webapp/requirements.txt
Flask==1.0.2
tensorflow
tensorflow-serving-api
python-mnist
Pillow
gunicorn
Flask-Cors==3.0.6